결재승인 insert 및 조회 + JPA prefix auto_increment 완성

+++ src/main/java/com/wisehr/wisehr/payment/entity/PaymentMember.java
@@ -15,7 +15,7 @@ public class PaymentMember {
     @Id
     @Column(name = "mem_code", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int memCode;
+    private Long memCode;
     @Column(name = "mem_name")
     private String memName;
     @Column(name = "mem_phone")
